chore: rename quantel/interfaces to quantel/types
