+++ packages/timeline-state-resolver/src/integrations/quantel/connection.ts
@@ -14,7 +14,7 @@ import {
 	QuantelStatePortClipContent,
 	QuantelTrackedState,
 	QuantelTrackedStatePort,
-} from './interfaces'
+} from './types'
 
 const SOFT_JUMP_WAIT_TIME = 250
