+++ packages/timeline-state-resolver/src/integrations/quantel/types.ts
